leaks pos heredoc

+++ redirects.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   redirects.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emsoares <emsoares@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jomirand <jomirand@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 15:23:20 by jomirand          #+#    #+#             */
-/*   Updated: 2023/07/12 11:27:42 by emsoares         ###   ########.fr       */
+/*   Updated: 2023/07/14 14:21:16 by jomirand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,13 @@ char	**handle_redirects(t_minishell *shell, char *command)
 	{
 		if(string_comp(command_args[i], ">"))
 		{
-			redirect_output(i, command_args);
+			if(redirect_output(i, command_args))
+			{
+				free_splited(command_args);
+				free_struct(shell);
+				free(command);
+				exit(2);
+			}
 			if(i == num_words - 2)
 			{
 				free(command_args[i]);
@@ -59,10 +65,29 @@ char	**handle_redirects(t_minishell *shell, char *command)
 		}
 		if(string_comp(command_args[i], "<<"))
 		{
-			here_doc(command_args[i + 1]);
+			if(here_doc(command_args[i + 1]))
+			{
+				free_splited(command_args);
+				free_struct(shell);
+				free(command);
+				exit(2);
+			}
+			if(i == num_words - 2 && i == 0)
+			{
+				free(command_args[i]);
+				free(command_args[i + 1]);
+				command_args[i] = 0;
+				return(command_args);
+			}
 			shell->stdin_fd = open(".heredoc", O_RDONLY);
-			//if(shell->stdin_fd < 0)
-				//handle_error
+			if (shell->stdin_fd < 0)
+			{
+				ft_putstr_fd("Minishell: no file specified in redirect '<<'.\n", 2);
+				free_splited(command_args);
+				free_struct(shell);
+				free(command);
+				exit(2);
+			}
 			unlink(".heredoc");
 		}
 		i++;
@@ -71,7 +96,7 @@ char	**handle_redirects(t_minishell *shell, char *command)
 	return(command_args);
 }
 
-void	redirect_output(int i, char **args)
+int	redirect_output(int i, char **args)
 {
 	int		file_fd;
 	char	*temp;
@@ -83,49 +108,86 @@ void	redirect_output(int i, char **args)
 		args[i + 1] = ft_strdup(temp);
 	}
 	file_fd = open(args[i + 1], O_WRONLY | O_CREAT | O_TRUNC, 0777);
+	if(!file_fd)
+	{
+		ft_putstr_fd("Minishell: no file specified in redirect '>'.\n", 2);
+		return(1);
+	}
 	dup2(file_fd, STDOUT_FILENO);
 	close(file_fd);
+	return(0);
 }
 
-void	redirect_input(int i, char **args)
+int	redirect_input(int i, char **args)
 {
 	int	file_fd;
+	char	*temp;
 
+	if(check_closed_quotes(args[i + 1]) == 1)
+	{
+		temp = quote_remover(args[i + 1]);
+		free(args[i + 1]);
+		args[i + 1] = ft_strdup(temp);
+	}
 	file_fd = open(args[i + 1], O_RDONLY);
+	if(!file_fd)
+	{
+		ft_putstr_fd("Minishell: no file specified in redirect '<'.\n", 2);
+		return(1);
+	}
 	dup2(file_fd, STDIN_FILENO);
 	close(file_fd);
+	return(0);
 }
 
-void	redirect_append(int i, char **args)
+int	redirect_append(int i, char **args)
 {
 	int	file_fd;
+	char	*temp;
 
+	if(check_closed_quotes(args[i + 1]) == 1)
+	{
+		temp = quote_remover(args[i + 1]);
+		free(args[i + 1]);
+		args[i + 1] = ft_strdup(temp);
+	}
 	file_fd = open(args[i + 1], O_WRONLY | O_CREAT | O_APPEND, 0777);
+	if(!file_fd)
+	{
+		ft_putstr_fd("Minishell: no file specified in redirect '>>'.\n", 2);
+		return(1);
+	}
 	dup2(file_fd, STDOUT_FILENO);
 	close(file_fd);
+	return(0);
 }
 
-void	here_doc(char *delimiter)
+int	here_doc(char *delimiter)
 {
 	int		temp_file;
 	char	*input;
 
 	temp_file = open(".heredoc", O_CREAT | O_WRONLY | O_TRUNC, S_IWUSR | S_IRUSR);
-	//if(temp_file < 0)
-		//handle_error
+	if(temp_file < 0)
+	{
+		ft_putstr_fd("Minishell: could not open .heredoc file\n", 2);
+		return(1);
+	}
 	while(1)
 	{
 		input = readline("> ");
-		//if(!input)
-			//handle_error
+		if(!input)
+		{
+			ft_putstr_fd("Minishell: heredoc interrupted by end-of-file\n", 2);
+			return(1);
+		}
 		if(string_comp(input, delimiter))
 			break ;
-		//if(ft_strrchr(input, '$'))
-			//input = expander
 		ft_putstr_fd(input, temp_file);
 		ft_putchar_fd('\n', temp_file);
 		free(input);
 	}
 	free(input);
 	close(temp_file);
+	return(0);
 }
